MODEXPS-1::Create mod-data-export-spring and mod-data-export-worker modules basing on the PoC Clean up REST API.
