+++ src/main/java/org/folio/des/controller/ConfigsController.java
@@ -1,22 +1,18 @@
 package org.folio.des.controller;
 
 import lombok.RequiredArgsConstructor;
-import lombok.extern.log4j.Log4j2;
 import org.folio.des.domain.dto.ExportConfig;
 import org.folio.des.domain.dto.ExportConfigCollection;
 import org.folio.des.rest.resource.ConfigsApi;
 import org.folio.des.service.ExportConfigService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.validation.Valid;
 
-@Log4j2
-@RequiredArgsConstructor
 @RestController
-@RequestMapping(value = "/")
+@RequiredArgsConstructor
 public class ConfigsController implements ConfigsApi {
 
   private final ExportConfigService configService;
