+++ src/main/resources/application.properties
@@ -1,5 +1,17 @@
-server.port=8081
 spring.application.name=mod-data-export-spring
+
+server.port=8081
+
+folio.tenant.validation.enabled=true
+
+spring.jackson.default-property-inclusion=NON_EMPTY
+
+spring.jpa.hibernate.ddl-auto=none
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.format_sql=true
+
+spring.liquibase.enabled=true
+spring.liquibase.changeLog=classpath:db/changelog/changelog-master.xml
 
 spring.kafka.bootstrap-servers=${KAFKA_HOST:localhost}:${KAFKA_PORT:9092}
 spring.kafka.producer.properties.interceptor.classes=org.folio.des.config.OkapiProducerInterceptor
@@ -10,6 +22,4 @@ spring.kafka.consumer.value-deserializer=org.apache.kafka.common.serialization.S
 spring.kafka.producer.key-serializer=org.apache.kafka.common.serialization.StringSerializer
 spring.kafka.producer.value-serializer=org.springframework.kafka.support.serializer.JsonSerializer
 
-folio.tenant.validation.enabled=true
-spring.liquibase.enabled=true
-spring.liquibase.changeLog=classpath:db/changelog/changelog-master.xml
+management.endpoints.web.exposure.include=info,health,env,httptrace,liquibase
