+++ src/main/resources/swagger.api/schemas/bursarFeeFines.json
@@ -18,6 +18,10 @@
           "type": "string"
         }
       }
-    }
+    },
+    "required": [
+      "daysOutstanding",
+      "patronGroups"
+    ]
   }
 }
