+++ src/main/resources/swagger.api/schemas/job.json
@@ -67,6 +67,9 @@
       }
     },
     "additionalProperties": false,
-    "required": ["type"]
+    "required": [
+      "type",
+      "exportTypeSpecificParameters"
+    ]
   }
 }
